making night tests small
